% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{cjet}
\alias{cjet}
\title{Create a jet colormap for use in e.g. heatmaps}
\usage{
cjet()
}
\value{
vector of hex color defining the jet colormap
}
\description{
Create a jet colormap for use in e.g. heatmaps
}
