% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{clincol}
\alias{clincol}
\title{Create a linear colormap for use in e.g. heatmaps}
\usage{
clincol()
}
\value{
vector of hex color defining the linear colormap
}
\description{
Create a linear colormap for use in e.g. heatmaps
}
